Revert "fix: Resolving bugs and change some things."
