+++ src/app/usuario/[id]/components/fields.tsx
@@ -7,6 +7,8 @@ import { Button } from '@/components/ui/button';
 import { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import axios from 'axios';
 import { useSession } from 'next-auth/react'
+import { api, fetchFromAPI } from '@/services/api'
+import { useFetch } from '@/hooks/useFetch'
 
 interface User {
   name: string,
@@ -111,10 +113,6 @@ const FormProfile = ({ params }: { params: { id: string } }) => {
 
   if (!user) {
     return <p>Usuário não encontrado</p>
-  }
-
-  if (!session) {
-    return <p>Você não está autorizado a visualizar!</p>;
   }
 
   return (
